refactor(settings): extract safe object helpers into src/lib
Move isSafePlainObject, deepMerge, and setNestedValue out of the
SettingsContext provider into a dedicated, dependency-free
src/lib/settingsObject.ts module, and test them directly there rather
than exporting provider internals purely for tests.

SettingsContext.test.tsx keeps the provider-behavior tests and no longer
imports the helpers. Coverage stays at 100% for both files.
